core_implementation/requirements.txt
langchain
langchain-groq
langchain-ollama
langchain-community
subprocess-tee
python-dotenv
PyPDF2
pyautogui
keyboard
groq
google-genai
google-generativeai
mss
sounddevice
soundfile
numpy
pyperclip
requests
beautifulsoup4
pandas
openpyxl
youtube-transcript-api
fpdf2
opencv-python
pytesseract
fastapi
uvicorn
psutil
networkx
pydantic
llama-cpp-python
pydantic-settings
